[mihome] Made constants and handler packages internal
