+++ addons/binding/org.openhab.binding.mihome/src/main/java/org/openhab/binding/mihome/internal/XiaomiHandlerFactory.java
@@ -8,7 +8,7 @@
  */
 package org.openhab.binding.mihome.internal;
 
-import static org.openhab.binding.mihome.XiaomiGatewayBindingConstants.*;
+import static org.openhab.binding.mihome.internal.XiaomiGatewayBindingConstants.*;
 
 import java.util.HashMap;
 import java.util.Hashtable;
@@ -24,26 +24,26 @@ import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
-import org.openhab.binding.mihome.handler.XiaomiActorCurtainHandler;
-import org.openhab.binding.mihome.handler.XiaomiActorGatewayHandler;
-import org.openhab.binding.mihome.handler.XiaomiActorPlugHandler;
-import org.openhab.binding.mihome.handler.XiaomiAqaraActorSwitch1Handler;
-import org.openhab.binding.mihome.handler.XiaomiAqaraActorSwitch2Handler;
-import org.openhab.binding.mihome.handler.XiaomiAqaraSensorSwitch1Handler;
-import org.openhab.binding.mihome.handler.XiaomiAqaraSensorSwitch2Handler;
-import org.openhab.binding.mihome.handler.XiaomiBridgeHandler;
-import org.openhab.binding.mihome.handler.XiaomiDeviceBaseHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorCubeHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorGasHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorHtHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorLockHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorMagnetHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorMotionHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorSmokeHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorSwitchHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorVibrationHandler;
-import org.openhab.binding.mihome.handler.XiaomiSensorWaterHandler;
 import org.openhab.binding.mihome.internal.discovery.XiaomiItemDiscoveryService;
+import org.openhab.binding.mihome.internal.handler.XiaomiActorCurtainHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiActorGatewayHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiActorPlugHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiAqaraActorSwitch1Handler;
+import org.openhab.binding.mihome.internal.handler.XiaomiAqaraActorSwitch2Handler;
+import org.openhab.binding.mihome.internal.handler.XiaomiAqaraSensorSwitch1Handler;
+import org.openhab.binding.mihome.internal.handler.XiaomiAqaraSensorSwitch2Handler;
+import org.openhab.binding.mihome.internal.handler.XiaomiBridgeHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiDeviceBaseHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorCubeHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorGasHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorHtHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorLockHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorMagnetHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorMotionHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorSmokeHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorSwitchHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorVibrationHandler;
+import org.openhab.binding.mihome.internal.handler.XiaomiSensorWaterHandler;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.annotations.Component;
 
@@ -106,45 +106,45 @@ public class XiaomiHandlerFactory extends BaseThingHandlerFactory {
     protected ThingHandler createHandler(Thing thing) {
         ThingTypeUID thingTypeUID = thing.getThingTypeUID();
 
-        if (thingTypeUID.equals(THING_TYPE_BRIDGE)) {
+        if (THING_TYPE_BRIDGE.equals(thingTypeUID)) {
             XiaomiBridgeHandler handler = new XiaomiBridgeHandler((Bridge) thing);
             registerItemDiscoveryService(handler);
             return handler;
-        } else if (thingTypeUID.equals(THING_TYPE_GATEWAY)) {
+        } else if (THING_TYPE_GATEWAY.equals(thingTypeUID)) {
             return new XiaomiActorGatewayHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_HT)) {
+        } else if (THING_TYPE_SENSOR_HT.equals(thingTypeUID)) {
             return new XiaomiSensorHtHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_MOTION)) {
+        } else if (THING_TYPE_SENSOR_MOTION.equals(thingTypeUID)) {
             return new XiaomiSensorMotionHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_SWITCH)) {
+        } else if (THING_TYPE_SENSOR_SWITCH.equals(thingTypeUID)) {
             return new XiaomiSensorSwitchHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_AQARA_SWITCH)) {
+        } else if (THING_TYPE_SENSOR_AQARA_SWITCH.equals(thingTypeUID)) {
             return new XiaomiSensorSwitchHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_MAGNET)) {
+        } else if (THING_TYPE_SENSOR_MAGNET.equals(thingTypeUID)) {
             return new XiaomiSensorMagnetHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_CUBE)) {
+        } else if (THING_TYPE_SENSOR_CUBE.equals(thingTypeUID)) {
             return new XiaomiSensorCubeHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_SMOKE)) {
+        } else if (THING_TYPE_SENSOR_SMOKE.equals(thingTypeUID)) {
             return new XiaomiSensorSmokeHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_GAS)) {
+        } else if (THING_TYPE_SENSOR_GAS.equals(thingTypeUID)) {
             return new XiaomiSensorGasHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_WATER)) {
+        } else if (THING_TYPE_SENSOR_WATER.equals(thingTypeUID)) {
             return new XiaomiSensorWaterHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_AQARA1)) {
+        } else if (THING_TYPE_SENSOR_AQARA1.equals(thingTypeUID)) {
             return new XiaomiAqaraSensorSwitch1Handler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_SENSOR_AQARA2)) {
+        } else if (THING_TYPE_SENSOR_AQARA2.equals(thingTypeUID)) {
             return new XiaomiAqaraSensorSwitch2Handler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_ACTOR_AQARA1)) {
+        } else if (THING_TYPE_ACTOR_AQARA1.equals(thingTypeUID)) {
             return new XiaomiAqaraActorSwitch1Handler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_ACTOR_AQARA2)) {
+        } else if (THING_TYPE_ACTOR_AQARA2.equals(thingTypeUID)) {
             return new XiaomiAqaraActorSwitch2Handler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_ACTOR_PLUG)) {
+        } else if (THING_TYPE_ACTOR_PLUG.equals(thingTypeUID)) {
             return new XiaomiActorPlugHandler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_ACTOR_AQARA_ZERO1)) {
+        } else if (THING_TYPE_ACTOR_AQARA_ZERO1.equals(thingTypeUID)) {
             return new XiaomiAqaraActorSwitch1Handler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_ACTOR_AQARA_ZERO2)) {
+        } else if (THING_TYPE_ACTOR_AQARA_ZERO2.equals(thingTypeUID)) {
             return new XiaomiAqaraActorSwitch2Handler(thing);
-        } else if (thingTypeUID.equals(THING_TYPE_ACTOR_CURTAIN)) {
+        } else if (THING_TYPE_ACTOR_CURTAIN.equals(thingTypeUID)) {
             return new XiaomiActorCurtainHandler(thing);
         } else if (THING_TYPE_SENSOR_AQARA_WEATHER_V1.equals(thingTypeUID)) {
             return new XiaomiSensorHtHandler(thing);
