+++ addons/binding/org.openhab.binding.mihome/src/main/java/org/openhab/binding/mihome/internal/handler/XiaomiSensorSwitchHandler.java
@@ -6,9 +6,9 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.mihome.handler;
+package org.openhab.binding.mihome.internal.handler;
 
-import static org.openhab.binding.mihome.XiaomiGatewayBindingConstants.CHANNEL_BUTTON;
+import static org.openhab.binding.mihome.internal.XiaomiGatewayBindingConstants.CHANNEL_BUTTON;
 
 import org.eclipse.smarthome.core.thing.Thing;
 import org.openhab.binding.mihome.internal.ChannelMapper;
